Refletindo a mesma estrutura de pastas do projeto Remoção de stubs
