--- tests/Unit/Core/Controller/_stub/Page/Abstract.php
@@ -1,56 +1,0 @@
-<?php
-
-/**
- * i-Educar - Sistema de gestão escolar
- *
- * Copyright (C) 2006  Prefeitura Municipal de Itajaí
- *                     <ctima@itajai.sc.gov.br>
- *
- * Este programa é software livre; você pode redistribuí-lo e/ou modificá-lo
- * sob os termos da Licença Pública Geral GNU conforme publicada pela Free
- * Software Foundation; tanto a versão 2 da Licença, como (a seu critério)
- * qualquer versão posterior.
- *
- * Este programa é distribuí­do na expectativa de que seja útil, porém, SEM
- * NENHUMA GARANTIA; nem mesmo a garantia implí­cita de COMERCIABILIDADE OU
- * ADEQUAÇÃO A UMA FINALIDADE ESPECÍFICA. Consulte a Licença Pública Geral
- * do GNU para mais detalhes.
- *
- * Você deve ter recebido uma cópia da Licença Pública Geral do GNU junto
- * com este programa; se não, escreva para a Free Software Foundation, Inc., no
- * endereço 59 Temple Street, Suite 330, Boston, MA 02111-1307 USA.
- *
- * @author      Eriksen Costa Paixão <eriksen.paixao_bs@cobra.com.br>
- * @category    i-Educar
- * @license     @@license@@
- * @package     Core_Controller
- * @subpackage  UnitTests
- * @since       Arquivo disponível desde a versão 1.1.0
- * @version     $Id$
- */
-
-require_once __DIR__.'/Abstract.php';
-require_once __DIR__.'/../../../../CoreExt/_stub/EntityDataMapper.php';
-
-/**
- * Core_Controller_Page_AbstractStub class.
- *
- * Stub para testes da class Core_Controller_Page_Abstract. Muda algumas
- * variáveis de instância para public, afim de fazer testes de inicialização
- * automática de valores.
- *
- * @author      Eriksen Costa Paixão <eriksen.paixao_bs@cobra.com.br>
- * @category    i-Educar
- * @license     @@license@@
- * @package     Core_Controller
- * @subpackage  UnitTests
- * @see         Core_Controller_Page_Abstract
- * @since       Classe disponível desde a versão 1.1.0
- * @version     @@package_version@@
- */
-class Core_Controller_Page_AbstractStub extends Core_Controller_Page_Abstract
-{
-  public $_dataMapper = NULL;
-  public $_processoAp = NULL;
-  public $_titulo = NULL;
-}
