--- tests/Unit/Core/_stub/View.php
@@ -1,58 +1,0 @@
-<?php
-
-/**
- * i-Educar - Sistema de gestão escolar
- *
- * Copyright (C) 2006  Prefeitura Municipal de Itajaí
- *                     <ctima@itajai.sc.gov.br>
- *
- * Este programa é software livre; você pode redistribuí-lo e/ou modificá-lo
- * sob os termos da Licença Pública Geral GNU conforme publicada pela Free
- * Software Foundation; tanto a versão 2 da Licença, como (a seu critério)
- * qualquer versão posterior.
- *
- * Este programa é distribuí­do na expectativa de que seja útil, porém, SEM
- * NENHUMA GARANTIA; nem mesmo a garantia implí­cita de COMERCIABILIDADE OU
- * ADEQUAÇÃO A UMA FINALIDADE ESPECÍFICA. Consulte a Licença Pública Geral
- * do GNU para mais detalhes.
- *
- * Você deve ter recebido uma cópia da Licença Pública Geral do GNU junto
- * com este programa; se não, escreva para a Free Software Foundation, Inc., no
- * endereço 59 Temple Street, Suite 330, Boston, MA 02111-1307 USA.
- *
- * @author      Eriksen Costa Paixão <eriksen.paixao_bs@cobra.com.br>
- * @category    i-Educar
- * @license     @@license@@
- * @package     Core_View
- * @subpackage  UnitTests
- * @since       Arquivo disponível desde a versão 1.1.0
- * @version     $Id$
- */
-
-require_once 'Core/View.php';
-
-/**
- * Core_ViewStub class.
- *
- * Stub para Core_View, substituindo alguns métodos para evitar conexão com
- * o banco de dados e a geração de código HTML.
- *
- * @author      Eriksen Costa Paixão <eriksen.paixao_bs@cobra.com.br>
- * @category    i-Educar
- * @license     @@license@@
- * @package     Core_View
- * @subpackage  UnitTests
- * @since       Classe disponível desde a versão 1.1.0
- * @version     @@package_version@@
- */
-class Core_ViewStub extends Core_View
-{
-  /**
-   * @see clsBase#MakeAll()
-   */
-  public function MakeAll()
-  {
-    $this->Formular();
-    return TRUE;
-  }
-}
