+++ tests/Unit/ieducar/intranet/ClsConfigItajaiTest.php
